requirements.txt
selenium>=4.0.0
pillow>=9.0.0
numpy>=1.21.0
urllib3>=2.0.0
webdriver_manager>=3.8.0